feat: Add createQuiz function to apiQuiz.ts
This commit adds a new function called createQuiz to the apiQuiz.ts file. The function is responsible for creating a new quiz in the database. It generates a unique quiz ID using the uuidv4 library and inserts the quiz details into the "quiz" table using the supabase client. The function returns the created quiz data if successful, or null if there was an error.